Renamed file and fixed python app
