+++ stonks_backend/__pycache__/hello.cpython-39.pyc
