+++ stonks_backend/__pycache__/main.cpython-38.pyc
