+++ stonks_backend/__pycache__/main.cpython-39.pyc
