+++ stonks_backend/__pycache__/sciencestuff.cpython-39.pyc
