+++ stonks_backend/finder.py
