+++ stonks_backend/plots/AAL.png
