+++ stonks_backend/plots/AAL_2017-09-23_year.png
