+++ stonks_backend/plots/AAL_2018-02-07_year.png
