+++ stonks_backend/plots/AAPL_2021-01-08_y.png
