+++ stonks_backend/plots/TSLA_2021-01-08_y.png
