+++ stonks_backend/plots/TWTR_2021-01-08_y.png
