+++ stonks_backend/sciencestuff.py
