+++ stonks_backend/sp500data.cpp
