+++ stonks_backend/stonkanalysis.py
