+++ stonks_backend/stonkdata.py
