+++ stonks_backend/teslastonk.csv
